refactor(Cluster):actions own component
