+++ src/app/virtualmachines/clustercard/clustercard.component.html
@@ -19,7 +19,7 @@
 			<div class="col-3" style="margin: auto; text-align: center">
 				{{ cluster?.userlogin }}
 				<br />
-				{{ cluster?.launch_date | date : 'yyyy.MM.dd HH:mm' }}
+				{{ cluster?.launch_date | date: 'yyyy.MM.dd HH:mm' }}
 			</div>
 		</div>
 		<div
@@ -55,177 +55,10 @@
 				</div>
 			</div>
 		</div>
-		<div class="card-body p-x-1 py-h">
-			<div
-				class="row"
-				*ngIf="!(this.cluster.project_is_migrated_to_simple_vm || this.cluster.migrate_project_to_simple_vm)"
-			>
-				<div class="col-md-auto" *ngIf="cluster?.status === 'Running'">
-					<a
-						*ngIf="!show_connection_info"
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'showHowToConnectButton_' + cluster?.cluster_id"
-						(click)="show_connection_info = true"
-					>
-						How to connect ...</a
-					>
-
-					<a
-						*ngIf="show_connection_info"
-						class="font-weight-bold font-xs btn-block btn btn-secondary"
-						[id]="'hideHowToConnectButton_' + cluster?.cluster_id"
-						(click)="show_connection_info = false"
-					>
-						... Hide information</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticDELETING &&
-						cluster?.status === 'Running'
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'scale_up_' + cluster?.cluster_id"
-						(click)="showScaleModal(SCALE_UP)"
-					>
-						Scale-Up</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticDELETING &&
-						cluster?.status === 'Running'
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'scale_down_' + cluster?.cluster_id"
-						(click)="showScaleModal(SCALE_DOWN)"
-					>
-						Scale-Down</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticDELETING &&
-						cluster?.status === 'Running'
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'generate_password_' + cluster?.cluster_id"
-						(click)="showPasswordModal()"
-					>
-						Generate Password</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticDELETING &&
-						cluster?.status === 'Running'
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'stop_' + cluster?.cluster_id"
-						(click)="showStopModal()"
-						onClick="return false;"
-					>
-						Stop</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticNOT_FOUND &&
-						cluster?.status !== VirtualMachineStates.staticDELETING
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'rename_' + cluster?.cluster_id"
-						(click)="showRenameModal()"
-						onClick="return false;"
-					>
-						Rename</a
-					>
-				</div>
-				<div
-					class="col-md-auto"
-					*ngIf="
-						cluster?.status !== VirtualMachineStates.staticDELETED &&
-						cluster?.status !== VirtualMachineStates.staticDELETING
-					"
-				>
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-danger"
-						[id]="'delete_' + cluster?.cluster_id"
-						[attr.data-test-id]="'delete_cluster_' + cluster?.cluster_id"
-						(click)="showDeleteModal()"
-						onClick="return false;"
-					>
-						Delete</a
-					>
-				</div>
-				<div class="col-md-auto" *ngIf="cluster?.status === VirtualMachineStates.staticSHUTOFF">
-					<a
-						class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-						[id]="'restart_' + cluster?.cluster_id"
-						(click)="showResumeModal()"
-						onClick="return false;"
-					>
-						Start</a
-					>
-				</div>
-				<div *ngIf="cluster?.status === VirtualMachineStates.staticNOT_FOUND" class="alert alert-warning" role="alert">
-					Please contact <a href="mailto:{{ CLOUD_PORTAL_SUPPORT_MAIL }}">{{ CLOUD_PORTAL_SUPPORT_MAIL }}</a>
-				</div>
-			</div>
-			<div
-				class="row"
-				*ngIf="this.cluster.project_is_migrated_to_simple_vm || this.cluster.migrate_project_to_simple_vm"
-			>
-				<div class="col-md-auto">
-					<div class="alert alert-info">
-						You are not able to control this instance via this platform as it is part of a migrated project!
-					</div>
-				</div>
-			</div>
-		</div>
-		<div class="card-footer p-x-1 py-h" *ngIf="show_connection_info">
-			<div>
-				<div class="row" *ngFor="let mode of cluster?.master_instance?.modes">
-					<div class="col-md-1">
-						<span class="font-weight-bold font-xs btn-block text-muted">{{ mode.name }}:</span>
-					</div>
-					<div class="col-md-9">
-						<span class="font-weight-bold font-xs btn-block text-muted" [innerHTML]="mode.description"></span>
-					</div>
-
-					<div class="col-md-2" style="margin: auto" *ngIf="mode.copy_field && mode.copy_field !== ''">
-						<div
-							class="font-weight-bold font-xs btn-block text-muted btn btn-outline-secondary"
-							(click)="copyToClipboard(mode.copy_field)"
-							[closeDelay]="3000"
-							ngbTooltip="Copied!"
-							triggers="click:blur"
-						>
-							Copy command
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
+		<app-cluster-actions
+			[cluster]="cluster"
+			(startStatusLoop)="check_status_loop()"
+			(stopStatusLoop)="stopAllCheckStatusTimer()"
+		></app-cluster-actions>
 	</div>
 </div>
