+++ src/app/virtualmachines/clusters/clusterdetail/clusterdetail.component.html
@@ -87,19 +87,11 @@
 
 			<div class="card-body" *ngIf="cluster?.status !== 'Deleted' && !errorOnLoading">
 				<div class="card-title"><strong>Actions</strong></div>
-				<div class="row">
-					<div class="col-2">
-						<button
-							type="button"
-							class="btn btn-danger"
-							style="width: 100%"
-							id="deleteCluster"
-							(click)="verifyModal.show()"
-						>
-							<strong>Delete Cluster</strong>
-						</button>
-					</div>
-				</div>
+				<app-cluster-actions
+					[cluster]="cluster"
+					(startStatusLoop)="check_status_loop()"
+					(stopStatusLoop)="stopAllCheckStatusTimer()"
+				></app-cluster-actions>
 			</div>
 		</div>
 
@@ -109,6 +101,7 @@
 			<div class="card-body">
 				<div class="card-title">
 					<strong>Instance: {{ cluster?.master_instance.name }}</strong>
+					<app-vmstatus style="margin-left: 20px" [vm]="cluster?.master_instance"></app-vmstatus>
 				</div>
 				<app-virtualmachineinfo [virtualMachine]="cluster?.master_instance"></app-virtualmachineinfo>
 			</div>
@@ -119,6 +112,7 @@
 			<div class="card-body" *ngFor="let worker of cluster?.worker_instances">
 				<div class="card-title">
 					<strong>Instance: {{ worker.name }}</strong>
+					<app-vmstatus style="margin-left: 20px" [vm]="worker"></app-vmstatus>
 				</div>
 				<app-virtualmachineinfo [cluster_machine]="true" [virtualMachine]="worker"></app-virtualmachineinfo>
 				<div class="btn-group">
